requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.1
django-cors-headers==4.9.0
fastapi==0.128.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pypdf==6.6.0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
whitenoise==6.11.0
google-genai==1.3.0
